rename components and setting sections
