+++ src/providers/options.tsx
@@ -68,37 +68,37 @@ export const options: Option[] = [
   {
     icon: <CurrencyIcon />,
     option: SettingsOptions.Currency,
-    section: SettingsSections.Display,
+    section: SettingsSections.General,
   },
   {
     icon: <CogIcon />,
     option: SettingsOptions.Display,
-    section: SettingsSections.Display,
+    section: SettingsSections.General,
   },
   {
     icon: <NotificationIcon />,
     option: SettingsOptions.Notifications,
-    section: SettingsSections.Display,
+    section: SettingsSections.General,
   },
   {
     icon: <NotesIcon />,
     option: SettingsOptions.Notes,
-    section: SettingsSections.Display,
+    section: SettingsSections.General,
   },
   {
     icon: <InfoIcon />,
     option: SettingsOptions.About,
-    section: SettingsSections.Display,
+    section: SettingsSections.General,
   },
   {
     icon: <SupportIcon />,
     option: SettingsOptions.Support,
-    section: SettingsSections.Display,
+    section: SettingsSections.General,
   },
   {
     icon: <PuzzleIcon />,
     option: SettingsOptions.Advanced,
-    section: SettingsSections.Display,
+    section: SettingsSections.General,
   },
   {
     icon: <BackupIcon />,
@@ -118,17 +118,17 @@ export const options: Option[] = [
   {
     icon: <></>,
     option: SettingsOptions.BitcoinUnit,
-    section: SettingsSections.Config,
+    section: SettingsSections.Display,
   },
   {
     icon: <></>,
     option: SettingsOptions.Haptics,
-    section: SettingsSections.Config,
+    section: SettingsSections.Display,
   },
   {
     icon: <></>,
     option: SettingsOptions.Theme,
-    section: SettingsSections.Config,
+    section: SettingsSections.Display,
   },
 ]
 
@@ -137,7 +137,7 @@ export interface SectionResponse {
   options: Option[]
 }
 
-const allOptions: SectionResponse[] = [SettingsSections.Display, SettingsSections.Security].map((section) => {
+const allOptions: SectionResponse[] = [SettingsSections.General, SettingsSections.Security].map((section) => {
   return {
     section,
     options: options.filter((o) => o.section === section),
@@ -172,14 +172,23 @@ export const OptionsProvider = ({ children }: { children: ReactNode }) => {
   const historyDepth = useRef(0)
 
   const optionSection = (opt: SettingsOptions): SettingsSections => {
-    return options.find((o) => o.option === opt)?.section || SettingsSections.Display
+    return options.find((o) => o.option === opt)?.section || SettingsSections.General
   }
 
   const getParentOption = (current: SettingsOptions): SettingsOptions => {
     const section = optionSection(current)
+    console.log('getParentOption', {
+      section,
+      parent:
+        section === SettingsSections.Advanced
+          ? SettingsOptions.Advanced
+          : section === SettingsSections.Display
+            ? SettingsOptions.Display
+            : SettingsOptions.Menu,
+    })
     return section === SettingsSections.Advanced
       ? SettingsOptions.Advanced
-      : section === SettingsSections.Config
+      : section === SettingsSections.Display
         ? SettingsOptions.Display
         : SettingsOptions.Menu
   }
